kyler I did it again
Added a variable to see if we are on Linux to make some Linux-specific changes to the code mainly so that it is less confusing if ppl on other operating systems are running Mac or something. Also made it so that it stops running if the OS is not Linux or Windows. Also added the console art.