Initial 2v2coopbattle Commit
The purpose of this commit is to set the foundry for this custom map specifically for this mod. This folder will contain any bsps used during development and a todo list as a progress tracker. This folder will be used as a workspace in the course of map development and will most likely be committed daily in order to keep a backup and for testing reasons with the mod.