Moving Files and Updating Them
I moved the workspace folder out of the ModFiles folder because the dev build would have been compiled/updated with all those files which it doesn't need. Fixed the .gitignore to target the right folders and files and added the .vmfs to store as a backup on GitHub in case I fuck up something.

Also with the hooks.nut file for downloading any custom files (line 208) looked incorrect so I corrected it, now it should download the content correctly. I also added aliasing for easy loading of the two custom maps 2v2coopbattle and p32lobby. Command for p32lobby will fail since it isn't fully built or implemented yet.