LOTS OF WORK
Started working on the save system. Trying to implement a system that map makers and us ourselves can use with custom maps or official maps to have some kind of saved progress.

This system will consist as a .nut and a .py each will talk to one another back and forth.

Added and changed a few things to get it compatible or make a place for it in the mod.