ADDING CUSTOM BUTTON STUFF
So I've got the ability built-in to change the position and size of buttons individually so they aren't always locked onto the side. The one issue though is having a linear scale for sizing and moving the buttons around. Also there is the fact the "\n" doesn't work and I have no idea why its not.  I have some text at the bottom left with xpos, ypos, and size defined, except 1. it goes off the screen, and 2. the position of the text also still corresponds to the position of the menu buttons, so if you want to have the Back button as seen you need to add it manually before you reference the buttons with custom position or size. I'm gonna try to make a individual button and text class so we have buttons that aren't reliant to the position of the menu, and text for a simple message that can be placed on the side or different places. Ignore the text saying top right, i was gonna use that to as a base to make the separate button class.