TONS OF SHIT DOOOOONE
First, I put the README.md through Grammarly and added \n and I to the Team Members list.

With cli.py, all I did was add a space in one place. Why do we still have this if it's gathering dust?

I have added the translations for the Save System things in English and French translations. I will need someone to check my translations in French to see if they match correctly.

In MainWindow, I added the functionality for the Save System. The menu will be a dev-only menu like the test menu. I also added DisplayText, a class similar to ButtonTemplate put the main point is to have a body of text that can wrap and fit into empty spaces in the menu. We can use it as instructions on the side or display something without needing to be tied to the menu layout or requiring a beep sound when hovering over it. An example of it in work is in the Saves menu. Size and movement are not linear, and DisplayText needs to be fully completed. It still needs its wrap functionally and needs to stick to the defined menu. Cabiste offered to finish this for me because he plans to optimize the code again. I also added Steam Deck support. I have yet to test it, though, so it will likely not work.

In config.nut I added \n's and mine IDs for admin in P2MM lobbies.

Ok, so with savesystem-main.nut, it still needs to be fully finished. I still need to do a lot. I have made sure it is not loaded in when a map starts. This system will take advantage of the map support system, but I still need to implement that fully. It will be able to check for .nut files created by SaveSystem.py and create .dem files, which SaveSystem.py will check for making a way to communicate between the launcher and Portal 2.

In GlobalVariables.py, I added stuff for the Steam Deck. Again, I do not know if it will work, but the functionality is added to detect it. I did need to import the platform library so it could determine its Steam OS from any other Linux build.

SaveSystem.py is fully functioning. It can boot up into the menu and display whether it has been enabled. It runs through if folders and specific files exist and does a check of the masterSave.cfg file, I need to optimize that check, though, with help from Cabiste, to ensure it is checking on what it needs to check. It also has custom errors to help pinpoint errors easier.

Broken Stuff or Fragmented Things:
There are still some things that need to be done or are in weird places or broken. The first thing is that the cube floaters are on the uppermost layer displaying in front of everything. I planned my DisplayText class to be in a separate list to show things from the button list because if I had the text in the button list, it would try to add button functionality and fail. The text also needs to wrap since the welcome text is off-screen. The DisplayText needs to also stick to the menus it's assigned to. Cabiste has volunteered to fix these things for me.