DisplayText and other text GUI optimizations, Steam Deck fixes, and FALLBACK removal
The stuff for DisplayText should be done as well as text wrapping for it. I also added text wrapping for InputPrompts so longer prompts can be used in the config, it will detect for "\n"s.

Made fixes for the Steam Deck, the p2mm folder will now be on the Desktop, it also should now actually start Portal 2. Changed the default launch window size width to 800 to accommodate the Steam Decks screen if the launcher is launched in Gaming Mode. One current bug though is that it is not pinging 8.8.8.8 to check if it has internet. Even if I am connected to the internet. I still need to continue test to determine and fix the issue.

Removed the FALLBACK stuff, it is not needed and is useless. The original point was to have another set of ModFiles in case the user isn't connected to the internet. You need internet to have users join you for P2MM. So its pointless, and so its removed.

Data System stuff is still not done yet, so I will be continuing to work on that, I just wanted to get my GUI changes for the launcher out before I continued the Data System.

The French translation needs a look over and fixed to accommodate a accurate French translations. Google Translate won't cut it.