README, Removal of File, Moving Of DRP stuff, and stuff for RunGame
Added to the README a shield badge for pypresence because the creator requests that for projects using pypresence.

Removed a file that was for a experiment.

Moved Discord Rich Presence stuff to its own script file because it will be tracking different things in the future and will needs it own space.

Added to RunGame a Log that will say what custom launch options are being passed to Portal 2 when it launches.