Fixed Discord RPC, Changed Names for GVars and DS
Fixed Discord RPC to properly detect if Discord is running or not.

Changed name for datasystemNutPath and value to instead of just targeting DS nut files to have scope of all of P2MM's nut files.

Moved and changed the name of the main datasystem file so its together with other main nut files. "datasystemsaves" will be where .nut files will be detected and created by the pythons Data System file.