Improved .nut removal at DS start, finally fixed DRP
Improved the old .nut file removal process to make sure all files despite their name and extension are removed clearing up the datasystemsaves folder.

Finally fixed Discord Rich Presence to work properly when Discord is not running and closing the connection when the launcher or Discord is shutdown.

Removed the counter from updateRichPresence to track if DRP was updating. Also added some logging info for exiting the P2MM launcher.