Properly fixed readme after accidental merge
The point of the previous merge was to merge the readme. The whole dev branch was accidentally merged when doing that. This change should fix the readme situation with what it should look like. Currently, the banner will use a temporary direct link to my fork since the main banner isn't currently on main.