Fixed logging messages and naming for things, also added more functions in datasystem.nut
Fixed logging messages so it is clear that this is the Data System reporting the information.

Fix the naming of some things like functions, data system indicators, and how a indicator is called apon.

Added stuff to set the password from the launcher and a request function so data can be retrieved from the masterData json.

All of this is still WIP and needs full testing instead of partial function testing.