Did a bunch of stuff for the Data System and Fixed Something In the Config
Moved some functions around and fixed some stuff up.

Fixed a bunch of comments and added a side logging system for the DS for specific things.

Worked on a almost finished password updater, still unfinished and still needs to change out the password in the listen server and server cfgs.

Added a checker to check for certain configs and do actions based on those changes.

The Data System is still in a unfinished state and has only been programmed to do stuff in the launcher, it will not do anything to Portal 2 yet.

Also fixed something in the config.