Finished DRP system for version 2.2
Finished DRP system for version 2.2, will be improved or fixed apon more for version 2.3.

Added checking if internet is available and if the user even wants to have DRP.

Added the ability to join the host via the button that appears in the hosts Discord profile. This will not appear for private servers that have a password.