A bunch of stuff for MainWindow
Fixed a lot of comments and added translations for a bunch of things that did not have a translation.

Did so much stuff for the Data System consisting of running the checker and refreshing it when a config change occurs.

Changed how ModFiles are found and the path that is used. DEVMOUNT ModFile mounting is currently broken though and I need help fixing :(

Added a function to check if both portal2_dlc1 and 2 exist, these are needed for P2MM to work properly.

Added some more instructional text on the menus to aid in what the menu is for and what it does.

Added ability to close the launcher using Ctrl/Command + C in the terminal window, this will help clean up the launcher when closing it instead of instantly killing it.