Readded Byte Patch and added check for Portal 2 DLC folders
Readded the byte patch for server passwords so at least there is some form of private server security. Added check for Portal 2's DLC folders because those contain content for multiplayer that we need. Removed a if-else statement for a config value that doesn't exist anymore, it was causing errors and isn't needed.