Updated ModFiles with latest dev
Updated the ModFiles with what was on dev branch by Nano, this includes Nano's latest work on the vscript, which I'm assuming he has made stable enough. I will though remove any features that are partially implemented or commented out to reduce size and to make sure nothing unstable or buggy is around in the next commit if need be.

Co-Authored-By: Nanoman2525 <75224575+Nanoman2525@users.noreply.github.com>