Steam Deck support (sort of) done
While all the essential stuff works for Steam Deck navigation in the launcher, two flaws that exist are that the shift on-screen keyboard button doesn't work/isn't detected, and the paste button doesn't paste, instead it inputs "v". Maybe there can be a work around for the pasting, maybe another button will suffice.