Optimized Mounting Encryption
The Encryption function didn't need its own python script so its been moved to RunGame and the function itself has been improved. Also added translations involving mounting to let the user know that mounting is occurring and that it is finished. Removed the Safe-Guard config, it's not used at all any where in code, so it's useless, and now it's gone.