Custom Lobby Music Control Has Been Completed!
I've finished the custom lobby music control fixing the issues that haunted me for several days.

One of those issues was clearing parenting, instead of clearing the parents for the buttons to move them to the prop arms, I cleared the prop arms and moved those to the buttons.

Some issues were caused due to timing within the MusicInit() function, those issues have been resolved by slightly delaying certain actions from clearing parents to moving entities.

Moved the user configs for the custom lobby music control to config.nut. There is a option to enable or disable the custom lobby music control, set the default music track, and where the end user can change which tracks are played.

Other miscellaneous bugs were resolved having to do with button skins as well as unlocking and locking of the music control buttons.

Also now by default, the DLC area door is open constantly, and the env_shakes and ambient_generic for the paint course reveal are gone because we have set it to already been revealed.