Updated the launcher to check for both identifier files
Updated the code in the launcher to check for both the old identifier file (32playermod.identifier) and the new one (p2mm.identifier) to maintain compatibility with older "ModFiles" builds.

While I've reamed the identifier file, I need to update the ModFile index which then also updates the date in the identifier file.