Updated the Updater to proper check currentVersion
Made a update to Updater.py to properly check that the currently used version, for both source code and compiled executable, is a version that is either newer or older than what's on the repository. It looks weird but allows when running the source code for it to behave more like it's a newer version, because normally it recognizes a dev build as a older build to what's been released already.

Also moved around other bits of code too to make it flow better.