VScript work, readded speedrun files, and removed spectator chat command code
Changed "[ERROR]" messages to have a explanation point at the end to make them feel more punchy. More a personal preference really.

Removed the Strip function, because it's useless.

Removed the cc file for spectator and the help command entry. It's not working so to save space its been removed.