Created new tool file updateModFiles.py
Created new tool file updateModFiles.py to update the p2mm/ModFiles from the src/ModFiles. Now before you say anything, yes this originally was a feature part of the launcher, but I had removed it originally because it was janky and broken, so now I've just made a separate script here.

Also added and edited some comments in the other updateModFIlesIndex.py.