Added team targeting in certain chat commands
Added the ability to target a specific team of players and replaced using "all" with `@a`. You can target Atlas team `@b`, P-Body team `@o`, and the Singleplayer team `@s`.

This was more of a QoL feature and also to make them similar to the targeting @'s of Minecraft. Like with Minecraft, the teleport chat command also has its arguments rearranged like the game. `!teleport (destination)` if its only one arg, but `!teleport (target) (destination)` when there are two arguments.

The team targeting operations have been applied to kill, rocket, slap, and teleport.

I've also updated the help sections accordingly with a small usage guide for each command when the player wants to see how to use a command when using `!help (command)`.