Removed map file, did a bunch of VScript work, and reduced how many logic_choreographed_scene entities get generated
Removed the mp_coop_lobby_3 map file. Originally this was here in case we ever were going to add support to other Portal 2 mods, so this was here as a way to load into that mod via lobby 3. We will most likely be adding support in the 3.0 series rather than the 2.0 series.

Changed a lot of while loops into for loops, just as a way to remove some unneeded lines of code and also because it makes more sense when read. They both functionally behave the same and have been tested to work.

Figured out a way to remove TONS of the logic_choreographed_scene entites that get created for glado's voicelines. It was due to the auto generated scenetable file not including checks for when a certain map even used certain scene files. This has ultimately brought down the logic_choreographed_scene entity count from a little over 700 entities to 277 entities. Should help prevent entity limit issues.