Small changes to the AppImage compile files
Added to the .dockerignore to ignore src/FALLBACK, we don't use it at all anymore.

Added the compiled AppImage files to the .gitignore.

Fixed the release version in AppImageBuilder.yml.

Made sure the old docker container in a previous AppImage compile is removed so a new compile can take place.