Updated the mp_coop_credits map support
Updated the mp_coop_credits map support to fix the last 40ish Valve credit entries from getting cut off.

I also took the liberty of updating our credits to fix up its formatting a bit as well as adding a new contributor and as well as who did what. Unfortunately, some of the entries I could not figure out a proper way to fit it all in the space that's provided.

I also moved the variable that stores the credits themselves from variable.nut to the map support file, that way we don't need a map check in variable.nut for mp_coop_credits.

Also made the P-Body that appears behind the monitor RAINBOW.