Consolidated and moved around menu items. Also updated some translations.
Consolidated the "Portal 2 Config" menu into the "Launcher Config" renaming "Launcher Config" to "General Settings". Also removed some unneeded Labels and moved the "Players" settings menu description label, now renamed to "Admins", into the "Admins" menu itself.

The "Encrypt-CVars" setting is removed because the old patching and encryption system is getting removed. This is because the plugin that's running in Portal 2 is now patching and encrypting at runtime. The next commit will involve the removal of the launcher's encrypting and patching system.

Updated and removed translations as needed. Apparently "update_client_text" was never translated for the other languages, it is now.