Removed the launcher's patching and encryption system and updated p2mm plugin
Removed the launcher's patching and encryption system because the newer version of the p2mm plugin does in at runtime in Portal 2. Now a bunch of code is removed from the launcher as it is not needed anymore.

Updated the p2mm plugin with the much newer and optimized version. This verison will be a preview of what 3.0's plugin will entail. This plugin will recieve several other updates before 2.2 releases to fix any lingering issues.

server.cfg has been removed as we won't be working with dedicated servers anymore. firstmapload.nut has been slightly modified to work with the new plugin.