Updated 2.2 version plugin
These edits are small and are simply there to work with the VScript for 2.2.

Made a edit to the log header to match the log header that the VScript does. Plugin -> PLUGIN

Changed the plugin version to indicate that this is a preview version of what is to come for the 3.0 rewrite.

Changed `p2mm_enable` to start as 2 in order to support the 2.2 VScript.

Disabled the ConVar which handles whether or not the `p2mm_welcome` message appears for a newly connected person. Just like Nano's comment says its currently a annoying and intrusive.