Adding check for system architecture and updated a toast
Added a check for what architecture CPU the system uses so we can install the right version of the launcher for systems. We use `platform.architecture` for WIndows and `platform.machine` for Linux. This isn't entirely finished though as there is a better way to do this that @cabiste69 will fully implement. This code will also make its way onto a patched release version of the launcher in preparation for 2.2 release as Linux now has the new AppImages.

Also updated a failed to update toast so it simply says that it failed and to go to the GitHub to download.