Moved mass entity removals below all necessary lobby fixes
On semi-rare occasion when a computer is slow and a loop of sorts is going multiple times, it will trigger the VScript VMs anti-constant loop measure killing the script that's doing the loop in order to save resources.

Moving the mass entity removals below the fixes ensures that the fixes are actually done before the entity removal so the lobby is fully playable. Problem before was that a loop was being killed so the fixes weren't able to be done.