Updated thread for Portal 2 and removal of p2mm dlc folder.
Originally if the launcher for some reason crashed or closed improperly while Portal 2 is running, Portal 2 would also close. Setting it as a daemon prevents this from happening so if for whatever reason it closes the P2:MM play session isn't interrupted.

If the launcher does close unexpectedly, the DLC folder doesn't get unmounted. Added a line which will remove any DLC folders that could still be lingering at startup.