Changed how some chat commands work
Changed the rocket chat command to be able to not need to rely on the loop in order to be killed after the launch. Also made it so non-admin players can use it to rocket themselves.

Same non-admin thing has been done with slap.

Changed "hostgunonly" to "duogunonly" because now it does only allows the first two players to have a portal gun to prevent softlocks in maps that need two players to have portal guns.