Sort of got Mel and Chell player models working
Got Mel to be set as the blue teams model, and Chell for the red team, but when the third player joins it doesn't work. Problem has to do with the player's PlayerID (their entindex) and related things with either PlayerID not being set. One of those things being their target name. Setting the model before without a delay had a previous issue also influenced by the target name not getting set where when a new client joins it sets the first players model also as the third clients target name hasn't been set properly.