Extras edits, and sp_a1_tramride cleanup
Replaced the sounds used when the splitscreen on and off buttons are pressed to give back feedback. Also had to change them because they came from PETI and PETI sounds aren't part of mel.

Cleaned up sp_a1_tramride a bit. The intro sequence now works properly, but any new players the join during the sequence won't be able to see it because the check for when players should be attached to the camera is broken. Will fix later.